\name{Baboons}
\alias{Baboons}
\docType{data}
\title{What time do the baboons come down from the trees?}
\description{
This data set arises in the course of analyzing data on the ecology of baboons in East Africa. The data consist on descent times of baboons (in hours since the day began) or
censoring times and the (left) censoring status.
}
\usage{data(Baboons)}
\format{
  A data frame with 152 observations on the following 2 variables.
  \describe{
    \item{\code{t}}{descent times of baboons or censoring times, in hours since the day began.}
    \item{\code{cs}}{(left) censoring status.}
  }
}
\references{
Wagner, S.S. and Altmann, S.A. (1973) What time do the baboons come down from the trees? (An estimation problem). Biometrics, 29: 623-635.}







