\name{influence.ssym}
\alias{influence.ssym}
\title{Tool to perform sensitivity analysis on the fitted model using local influence measures.}
\description{
\bold{influence} extracts from a object of class \code{ssym} the local influence measures and displays their graphs versus the index of the observations.}

\references{Cook, R.D. (1986). Assessment Local Influence (with discussion). Journal of the Royal Statistical Society Series B (Methodological). 48, 133-169.

			Poon, W.Y. and Poon, Y.S. (1999). Conformal Normal Curvature and Assessment of Local Influence. Journal of the Royal Statistical Society Series B (Methodological). 61, 51-61.}

\author{Luis Hernando Vanegas <hvanegasp@gmail.com> and Gilberto A. Paula}

