\name{plot.ssym}
\alias{plot.ssym}
\title{plot.ssym}
\description{
\bold{plot.ssym} produces the graph in which the overall goodness-of-fit statistic \eqn{\Upsilon} is based.
This function also displays graphs of the deviance-type residuals versus the fitted values for the median and the skewness (or the relative dispersion) submodels.
Under the presence of an uncensored sample, the function \code{plot()} produces a graph of the standardized individual-specific weights versus the ordinary residuals (i.e., a graph
of \eqn{\rho(\hat{{z}}_k)} versus \eqn{\hat{{z}}_k}, \eqn{k=1,\ldots,n}), and under the presence of a right-censored sample, the function \code{plot()} produces a graph
of the survival function of the error distribution.}
