\name{print.ssym}
\alias{print.ssym}
\title{print.ssym}
\description{
\bold{print.ssym} displays a summary (simpler than summary.ssym) of the fitted model including parameter estimates, associated (approximated) standard errors and
goodness-of-fit statistics from an object of class ``"ssym".}
