\name{Steel}
\alias{Steel}
\docType{data}
\title{Hardened Steel}
\description{
This dataset consists of the failure times for hardened steel specimens in a rolling contact fatigue test. Ten independent observations were
taken at each of the four values of contact stress. The response is the length of the time until each specimen of the hardened steel failed.
}
\usage{data(Steel)}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{stress}}{values of contact stress, in pounds per square inch x \eqn{10^{-6}}}
    \item{\code{life}}{length of the time until the specimen of the hardened steel failed.}
  }
}
\references{
McCool, J. (1980) Confidence limits for Weibull regression with censored data. Transactions on Reliability, 29: 145-150.}







